[ADD] implement login page with logic controller
